+++ plugins/google-analytics/index.php
@@ -3,12 +3,12 @@
 Plugin Name: Google Analytics
 Plugin URI: https://github.com/mindstellar/shopclass-plugins
 Description: Add Google Analytics 4 to your site, with consent gating and a per-browser opt-out.
-Version: 1.0.0
+Version: 1.0.1
 Author: Navjot Tomer (Mindstellar)
 Author URI: https://mindstellar.com
 Short Name: google-analytics
 Requires Shopclass: 6.1.0
-Tested up to: 6.2
+Tested up to: 6.4
 Requires PHP: 8.0
 Support URI: https://github.com/mindstellar/shopclass-plugins/issues
 */