+++ plugins/registered-users-only/index.php
@@ -3,12 +3,12 @@
 Plugin Name: Registered Users Only
 Plugin URI: https://github.com/mindstellar/shopclass-plugins
 Description: Require a visitor to sign in before they can see the site, with control over what stays public.
-Version: 2.0.0
+Version: 2.0.1
 Author: Navjot Tomer (Mindstellar)
 Author URI: https://mindstellar.com
 Short Name: registered-users-only
 Requires Shopclass: 6.1.0
-Tested up to: 6.2
+Tested up to: 6.4
 Requires PHP: 8.0
 Support URI: https://github.com/mindstellar/shopclass-plugins/issues
 */